refactor(tui): rename event.rs to event_bus.rs
