+++ src/tui/cleanup_app.rs
@@ -16,7 +16,7 @@ use ratatui::{
 };
 
 use super::app::App;
-use super::event::Event;
+use super::event_bus::Event;
 use super::preview_cache::PreviewCache;
 use super::widgets::{FileExplorer, FileExplorerWidget, FileItem};
 use crate::theme::current_theme;
