+++ src/tui/event_bus.rs
