+++ src/tui/mod.rs
@@ -8,7 +8,7 @@
 
 pub mod app;
 pub mod cleanup_app;
-pub mod event;
+pub mod event_bus;
 pub mod list_app;
 pub mod preview_cache;
 pub mod ui;
